+++ domain/dtos/result_dto.py
@@ -1,0 +1,9 @@
+from pydantic import BaseModel
+
+from domain.dtos.blob_stats_dto import BlobStatsDto
+
+
+class ResultDto(BaseModel):
+    blob: BlobStatsDto
+    position: int
+    points: int