+++ web-ui/src/components/pages/DashboardPage.tsx
@@ -1,5 +1,4 @@
 import { Box, Card, CardContent, Typography } from '@mui/material';
-import { useState } from 'react';
 import { PageFrame } from '../common/PageFrame';
 import { SimTimeDisplay } from '../common/SimTimeDisplay';
 import { useSimTime } from '../../context/SimTimeContext';
@@ -7,13 +6,12 @@ import { useTranslation } from 'react-i18next';
 import { NewsAndFooter } from '../dashboard-components/NewsAndFooter';
 
 export function DashboardPage() {
-  const [loadingOverlayVisible, setLoadingOverlayVisible] = useState(false);
   const { t } = useTranslation();
 
   const { loading: simTimeLoading } = useSimTime();
 
   return (
-    <PageFrame showLoading={loadingOverlayVisible || simTimeLoading} pageName="dashboard">
+    <PageFrame showLoading={simTimeLoading} pageName="dashboard">
       <Card>
         <CardContent>
           <Box display="flex" flexDirection="column" gap={1}>
@@ -22,7 +20,7 @@ export function DashboardPage() {
           </Box>
         </CardContent>
       </Card>
-      <NewsAndFooter setLoadingOverlayVisible={setLoadingOverlayVisible} />
+      <NewsAndFooter />
     </PageFrame>
   );
 }